small fixes

+++ src/ft_vec2i.c
@@ -6,7 +6,7 @@
 /*   By: myli-pen <myli-pen@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/18 14:32:13 by myli-pen          #+#    #+#             */
-/*   Updated: 2025/07/21 14:10:25 by myli-pen         ###   ########.fr       */
+/*   Updated: 2025/07/25 19:03:18 by myli-pen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,8 +25,8 @@ t_vec2i	vec2i_n(float n)
 {
 	t_vec2i	result;
 
-	result.x = n;
-	result.y = n;
+	result.x = (int)n;
+	result.y = (int)n;
 	return (result);
 }
 
@@ -52,7 +52,7 @@ t_vec2i	vec2i_scale(t_vec2i v, float s)
 {
 	t_vec2i	result;
 
-	result.x = (float)v.x * s;
-	result.y = (float)v.y * s;
+	result.x = (int)((float)v.x * s);
+	result.y = (int)((float)v.y * s);
 	return (result);
 }
